refactor: User 데코레이터 모듈 수정
